refactor(farm): isolate editor state and placement validation
