+++ app/page.tsx
@@ -1,10 +1,12 @@
 "use client";
 
+import { useFarmEditor } from "./dashboard/use-farm-editor";
+
 import { useFarmCanvas } from "./dashboard/use-farm-canvas";
 
 import { useDashboardNavigation } from "./dashboard/use-dashboard-navigation";
 
-import { reconcileProposals, buildingType, buildingSignature } from "./dashboard/farm-model";
+import { buildingType, buildingSignature } from "./dashboard/farm-model";
 
 import { useI18n } from "./i18n";
 import { useRef } from "react";
@@ -12,16 +14,15 @@ import { useState } from "react";
 import { useCallback } from "react";
 import { useEffect } from "react";
 import type { AppLanguageMode } from "./i18n";
-import { useMemo } from "react";
 import type { CSSProperties } from "react";
 import { ChangelogHistory } from "./changelog";
-import { type Snapshot, type LiveState, type Tile, type Suggestion, type Terrain, type StorageInventoryItem } from "./dashboard/snapshot-types";
+import { type Snapshot, type LiveState, type StorageInventoryItem } from "./dashboard/snapshot-types";
 import { type FarmHistory, type ActiveView, type SessionSummary, type LiveAlertSettings, type DesktopDiagnostics, type FarmOption, type UpdateState, type DesktopUpdates } from "./dashboard/ui-types";
 import { defaultLiveAlertSettings, deriveLiveAlerts, LiveDataPanel, LiveAlertCenter } from "./dashboard/live-view";
 import { localizedUpdateMessage, seasonName, localizedTerrainFeature, buildingDisplayName, communityBundleName, communityRoomName, buildingPlanText, buildingCategoryName, formatGameDate, formatLiveTime, localizedInteriorName } from "./dashboard/formatting";
 import { APPLICATION_VERSION, sessionSummary, liveStorageSource, feedbackIssueUrl } from "./dashboard/selectors";
 import { localizeSnapshotGameNames } from "./dashboard/game-names";
-import { spritePaths, tileKey, TILE, tools, BuildingPreview, InteriorView } from "./dashboard/farm-rendering";
+import { spritePaths, tileKey, tools, BuildingPreview, InteriorView } from "./dashboard/farm-rendering";
 import { ItemArtworkCatalogContext } from "./dashboard/artwork";
 import { LanguageModeIcon, Toggle } from "./dashboard/ui";
 import { ItemLocationDialog } from "./dashboard/storage";
@@ -36,22 +37,11 @@ export default function Home() {
   const appShellRef = useRef<HTMLElement>(null);
   const topbarRef = useRef<HTMLElement>(null);
   const [progressTabsTop, setProgressTabsTop] = useState(82);
-  const [initialMapPreferences] = useState<Record<string, unknown>>(() => {
-    if (typeof window === "undefined") return {};
-    try {
-      return JSON.parse(
-        window.localStorage.getItem("stardew-tool-map-preferences") ||
-          window.localStorage.getItem("aincrad-map-preferences") ||
-          "{}",
-      );
-    } catch {
-      return {};
-    }
-  });
-  const canvasRef = useRef<HTMLCanvasElement>(null);
+
+
   const workspaceRef = useRef<HTMLElement>(null);
-  const mapViewportRef = useRef<HTMLDivElement>(null);
-  const hasCenteredFarmRef = useRef(false);
+
+
   const [data, setData] = useState<Snapshot | null>(null);
   const [previousDay, setPreviousDay] = useState<Snapshot | null>(null);
   const [history, setHistory] = useState<FarmHistory>({
@@ -500,85 +490,31 @@ export default function Home() {
   }, [rightPanelWidth]);
   const [lastRefresh, setLastRefresh] = useState<Date | null>(null);
   const [live, setLive] = useState<LiveState>({ active: false });
+  const { mapData, selected, mapLocation, showState, setShowState, showProduction, setShowProduction, proposalStates, showSuggestions, setShowSuggestions, showGrid, setShowGrid, showBlocked, setShowBlocked, proposalEditMode, setProposalEditMode, setTool, setMovingProposalId, tool, proposalUndo, setProposalUndo, localSuggestions, persist, setMapLocation, setSelected, setPlacementError, centerOnFarmhouse, setZoom, zoom, mapViewportRef, canvasRef, setHover, pointFromEvent, handleClick, openProposalMenu, proposalMenu, setProposalMenu, placementError, persistProposalResolutions, proposalResolutions, persistProposalLinks, proposalLinks , hover, movingProposalId } = useFarmEditor(data, live, activeView);
   const [base, setBase] = useState<HTMLImageElement | null>(null);
   const [sprites, setSprites] = useState<Record<string, HTMLImageElement>>({});
   const [assetError, setAssetError] = useState("");
   const [dataLoadError, setDataLoadError] = useState("");
-  const [zoom, setZoom] = useState(() =>
-    typeof initialMapPreferences.zoom === "number"
-      ? Math.max(0.65, Math.min(2.1, initialMapPreferences.zoom))
-      : 1,
-  );
-  const [hover, setHover] = useState<Tile | null>(null);
-  const [selected, setSelected] = useState<Tile | null>(null);
-  const [tool, setTool] = useState("inspect");
-  const [proposalEditMode, setProposalEditMode] = useState(false);
-  const [movingProposalId, setMovingProposalId] = useState<string | null>(null);
-  const [proposalMenu, setProposalMenu] = useState<{
-    id: string;
-    name: string;
-    x: number;
-    y: number;
-  } | null>(null);
-  const [proposalUndo, setProposalUndo] = useState<Suggestion[] | null>(null);
-  const [showGrid, setShowGrid] = useState(() =>
-    typeof initialMapPreferences.showGrid === "boolean"
-      ? initialMapPreferences.showGrid
-      : false,
-  );
-  const [showState, setShowState] = useState(() =>
-    typeof initialMapPreferences.showState === "boolean"
-      ? initialMapPreferences.showState
-      : true,
-  );
-  const [showProduction, setShowProduction] = useState(() =>
-    typeof initialMapPreferences.showProduction === "boolean"
-      ? initialMapPreferences.showProduction
-      : true,
-  );
-  const [showBlocked, setShowBlocked] = useState(() =>
-    typeof initialMapPreferences.showBlocked === "boolean"
-      ? initialMapPreferences.showBlocked
-      : false,
-  );
-  const [showSuggestions, setShowSuggestions] = useState(() =>
-    typeof initialMapPreferences.showSuggestions === "boolean"
-      ? initialMapPreferences.showSuggestions
-      : true,
-  );
-  const [localSuggestions, setLocalSuggestions] = useState<Suggestion[]>([]);
-  const [proposalLinks, setProposalLinks] = useState<Record<string, string>>({});
-  const [proposalResolutions, setProposalResolutions] = useState<
-    Record<string, "resolved">
-  >({});
-  const [placementError, setPlacementError] = useState("");
-  const [mapLocation, setMapLocation] = useState(() =>
-    typeof initialMapPreferences.location === "string"
-      ? initialMapPreferences.location
-      : "farm",
-  );
 
-  useEffect(() => {
-    fetch("/api/preferences", { cache: "no-store" })
-      .then((response) => response.json())
-      .then((preferences) => {
-        if (Array.isArray(preferences.suggestions))
-          setLocalSuggestions(preferences.suggestions);
-        if (
-          preferences.proposalLinks &&
-          typeof preferences.proposalLinks === "object" &&
-          !Array.isArray(preferences.proposalLinks)
-        )
-          setProposalLinks(preferences.proposalLinks);
-        if (
-          preferences.proposalResolutions &&
-          typeof preferences.proposalResolutions === "object" &&
-          !Array.isArray(preferences.proposalResolutions)
-        )
-          setProposalResolutions(preferences.proposalResolutions);
-      })
-      .catch(() => undefined);
-  }, []);
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
   useEffect(() => {
     const expectedProfileId = data?.profileId;
@@ -622,28 +558,7 @@ export default function Home() {
     return () => window.clearInterval(timer);
   }, [data?.profileId]);
 
-  useEffect(() => {
-    window.localStorage.setItem(
-      "stardew-tool-map-preferences",
-      JSON.stringify({
-        zoom,
-        location: mapLocation,
-        showGrid,
-        showState,
-        showProduction,
-        showBlocked,
-        showSuggestions,
-      }),
-    );
-  }, [
-    mapLocation,
-    showBlocked,
-    showGrid,
-    showProduction,
-    showState,
-    showSuggestions,
-    zoom,
-  ]);
+
 
   useEffect(() => {
     let loadingLatest = false;
@@ -781,13 +696,7 @@ export default function Home() {
       .catch(() => setPreviousDay(null));
   }, [data, gameCatalog, history, t]);
 
-  useEffect(() => {
-    if (!data || mapLocation === "farm") return;
-    if (!data.interiors.some((interior) => interior.id === mapLocation)) {
-      const frame = window.requestAnimationFrame(() => setMapLocation("farm"));
-      return () => window.cancelAnimationFrame(frame);
-    }
-  }, [data, mapLocation]);
+
 
   useEffect(() => {
     if (!data?.dailyBrief) return;
@@ -801,259 +710,31 @@ export default function Home() {
     }
   }, [data]);
 
-  const persist = (next: Suggestion[], remember = true) => {
-    if (remember) setProposalUndo(localSuggestions);
-    setLocalSuggestions(next);
-    fetch("/api/preferences", {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ suggestions: next }),
-    }).catch(() => undefined);
-  };
 
-  const persistProposalLinks = (next: Record<string, string>) => {
-    setProposalLinks(next);
-    fetch("/api/preferences", {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ proposalLinks: next }),
-    }).catch(() => undefined);
-  };
 
-  const persistProposalResolutions = (next: Record<string, "resolved">) => {
-    setProposalResolutions(next);
-    fetch("/api/preferences", {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ proposalResolutions: next }),
-    }).catch(() => undefined);
-  };
 
-  const mapData = useMemo(() => {
-    if (!data || !live.active || !live.farmMap) return data;
-    const savedTerrain = new Map(
-      data.terrain.map((feature) => [tileKey(feature.x, feature.y), feature]),
-    );
-    const terrain = live.farmMap.terrain.map((feature) => {
-      const saved = savedTerrain.get(tileKey(feature.x, feature.y));
-      if (!saved)
-        return {
-          x: feature.x,
-          y: feature.y,
-          kind: feature.kind,
-          watered: feature.watered,
-        } as Terrain;
-      if (feature.kind === "HoeDirt" && !feature.hasCrop) {
-        const soil = { ...saved };
-        delete soil.crop;
-        delete soil.phase;
-        delete soil.cropRow;
-        return { ...soil, watered: feature.watered };
-      }
-      return { ...saved, watered: feature.watered };
-    });
-    return {
-      ...data,
-      terrain,
-      objects: live.farmMap.objects,
-      buildings: live.farmMap.buildings,
-    };
-  }, [data, live.active, live.farmMap]);
 
-  const proposalStates = useMemo(
-    () =>
-      mapData
-        ? reconcileProposals(
-            [...mapData.suggestions, ...localSuggestions],
-            mapData.buildings,
-            proposalLinks,
-            proposalResolutions,
-          )
-        : [],
-    [localSuggestions, mapData, proposalLinks, proposalResolutions],
-  );
 
-  const centerOnFarmhouse = useCallback(
-    (behavior: ScrollBehavior = "auto") => {
-      const viewport = mapViewportRef.current;
-      const farmhouse = mapData?.buildings.find(
-        (building) => buildingType(building) === "farmhouse",
-      );
-      if (!viewport || !farmhouse) return;
-      const centerX = (farmhouse.x + farmhouse.width / 2) * TILE * zoom;
-      const centerY = (farmhouse.y + farmhouse.height / 2) * TILE * zoom;
-      viewport.scrollTo({
-        left: Math.max(0, centerX - viewport.clientWidth / 2),
-        top: Math.max(0, centerY - viewport.clientHeight / 2),
-        behavior,
-      });
-    },
-    [mapData, zoom],
-  );
 
-  useEffect(() => {
-    if (activeView !== "map" || mapLocation !== "farm") return;
-    if (hasCenteredFarmRef.current) return;
-    if (
-      !mapData ||
-      !mapViewportRef.current ||
-      !mapData.buildings.some(
-        (building) => buildingType(building) === "farmhouse",
-      )
-    )
-      return;
-    hasCenteredFarmRef.current = true;
-    const frame = window.requestAnimationFrame(() => centerOnFarmhouse("auto"));
-    return () => window.cancelAnimationFrame(frame);
-  }, [activeView, centerOnFarmhouse, mapData, mapLocation]);
 
-  const validatePlacement = (point: Tile, width: number, height: number) => {
-    if (!mapData) return t("map.error.unavailable");
-    const cells = Array.from({ length: width * height }, (_, index) => ({
-      x: point.x + (index % width),
-      y: point.y + Math.floor(index / width),
-    }));
-    if (
-      cells.some(
-        (cell) =>
-          cell.x < 0 ||
-          cell.y < 0 ||
-          cell.x >= mapData.map.width ||
-          cell.y >= mapData.map.height,
-      )
-    )
-      return t("map.error.outsideFarm");
-    if (
-      cells.some((cell) =>
-        mapData.map.blocked.some(([x, y]) => x === cell.x && y === cell.y),
-      )
-    )
-      return t("map.error.nonBuildable");
-    if (
-      cells.some((cell) =>
-        mapData.buildings.some(
-          (building) =>
-            cell.x >= building.x &&
-            cell.x < building.x + building.width &&
-            cell.y >= building.y &&
-            cell.y < building.y + building.height,
-        ),
-      )
-    )
-      return t("map.error.existingBuilding");
-    if (
-      cells.some((cell) =>
-        mapData.objects.some(
-          (object) => {
-            if (object.x !== cell.x || object.y !== cell.y) return false;
-            if (object.kind === "Litter" || object.name === "Artifact Spot")
-              return false;
-            return !mapData.terrain.some(
-              (feature) =>
-                feature.x === object.x &&
-                feature.y === object.y &&
-                ["Tree", "FruitTree"].includes(feature.kind),
-            );
-          },
-        ),
-      )
-    )
-      return t("map.error.placedObject");
-    if (
-      cells.some((cell) =>
-        proposalStates.some(
-          (proposal) =>
-            proposal.status === "pending" &&
-            cell.x >= proposal.x &&
-            cell.x < proposal.x + proposal.width &&
-            cell.y >= proposal.y &&
-            cell.y < proposal.y + proposal.height,
-        ),
-      )
-    )
-      return t("map.error.pendingProposal");
-    return "";
-  };
+
+
+
+
+
+
+
+
 
   const draw = useFarmCanvas({ canvasRef, base, hover, mapData, localSuggestions, movingProposalId, proposalEditMode, proposalStates, showBlocked, showGrid, showProduction, showState, showSuggestions, sprites, tool, activeView, mapLocation });
 
-  const pointFromEvent = (event: React.MouseEvent<HTMLCanvasElement>) => {
-    const rect = event.currentTarget.getBoundingClientRect();
-    return {
-      x: Math.floor(((event.clientX - rect.left) / rect.width) * 80),
-      y: Math.floor(((event.clientY - rect.top) / rect.height) * 65),
-    };
-  };
 
-  const handleClick = (event: React.MouseEvent<HTMLCanvasElement>) => {
-    setProposalMenu(null);
-    const point = pointFromEvent(event);
-    setSelected(point);
-    if (movingProposalId) {
-      const moving = localSuggestions.find((item) => item.id === movingProposalId);
-      if (!moving) return setMovingProposalId(null);
-      const invalid = validatePlacement(point, moving.width, moving.height);
-      if (invalid) return setPlacementError(invalid);
-      persist(localSuggestions.map((item) =>
-        item.id === movingProposalId ? { ...item, x: point.x, y: point.y } : item,
-      ));
-      setMovingProposalId(null);
-      setPlacementError("");
-      return;
-    }
-    if (proposalEditMode && tool !== "inspect") {
-      const active = tools.find((item) => item.id === tool)!;
-      const invalid = validatePlacement(point, active.width, active.height);
-      if (invalid) {
-        setPlacementError(invalid);
-        return;
-      }
-      setPlacementError("");
-      persist([
-        ...localSuggestions,
-        {
-          id: `${tool}-${Date.now()}`,
-          kind: tool,
-          name: `Proposed ${active.label}`,
-          x: point.x,
-          y: point.y,
-          width: active.width,
-          height: active.height,
-          color: "#ffcf5c",
-        },
-      ]);
-      setTool("inspect");
-    }
-  };
 
-  const openProposalMenu = (event: React.MouseEvent<HTMLCanvasElement>) => {
-    event.preventDefault();
-    const point = pointFromEvent(event);
-    const proposal = [...localSuggestions].reverse().find(
-      (item) =>
-        point.x >= item.x &&
-        point.x < item.x + item.width &&
-        point.y >= item.y &&
-        point.y < item.y + item.height,
-    );
-    if (!proposal) {
-      setProposalMenu(null);
-      return;
-    }
-    setProposalMenu({
-      id: proposal.id,
-      name: proposal.name.replace(/^(Proposed|Future|Optional) /, ""),
-      x: event.clientX,
-      y: event.clientY,
-    });
-  };
 
-  useEffect(() => {
-    if (!proposalMenu) return;
-    const close = () => setProposalMenu(null);
-    document.addEventListener("click", close);
-    return () => document.removeEventListener("click", close);
-  }, [proposalMenu]);
+
+
+
+
 
   const beginPanelResize = (
     side: "left" | "right",
